extract TestHibernateQuery from ConcurrentCacheTest
